# api/swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "http"
  ],
  "swagger": "2.0",
  "info": {
    "description": "The purpose of this application is to be a batch processing system.",
    "title": "Arrebol api documentation",
    "license": {
      "name": "MIT",
      "url": "http://opensource.org/licenses/MIT"
    },
    "version": "0.0.1"
  },
  "host": "localhost:8080",
  "paths": {
    "/v1/queues/": {
      "get": {
        "description": "Retrieve current queues",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getQueues",
        "responses": {
          "200": {
            "description": "The current queues",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/QueueResponse"
              }
            }
          }
        }
      },
      "post": {
        "description": "Creates a queue",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "createQueue",
        "parameters": [
          {
            "description": "The queue payload",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Queue"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The queue ID",
            "schema": {
              "$ref": "#/definitions/GenericIdResponse"
            }
          }
        }
      }
    },
    "/v1/queues/{queue_id}": {
      "get": {
        "description": "Retrieve queue by its id",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "getQueue",
        "parameters": [
          {
            "type": "string",
            "description": "The queue id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "The queue",
            "schema": {
              "$ref": "#/definitions/QueueResponse"
            }
          }
        }
      }
    },
    "/v1/queues/{queue_id}/jobs": {
      "get": {
        "description": "Retrieve jobs by queue",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "retrieveJobsByQueue",
        "parameters": [
          {
            "type": "string",
            "description": "The queue id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The jobs",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Job"
              }
            }
          }
        }
      },
      "post": {
        "description": "Create a job",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "createJob",
        "parameters": [
          {
            "type": "string",
            "description": "The queue id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "description": "The job payload",
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/jobSpec"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "The job id",
            "schema": {
              "$ref": "#/definitions/GenericIdResponse"
            }
          }
        }
      }
    },
    "/v1/queues/{queue_id}/jobs/{job_id}": {
      "get": {
        "description": "Retrieve job by queue",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "operationId": "retrieveJobByQueue",
        "parameters": [
          {
            "type": "string",
            "description": "The queue id",
            "name": "id",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "description": "The job id",
            "name": "id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "The jobs",
            "schema": {
              "$ref": "#/definitions/Job"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Command": {
      "type": "object",
      "properties": {
        "Commands": {
          "type": "integer",
          "format": "int8",
          "x-go-name": "ExitCode"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "RawCommand": {
          "type": "string"
        },
        "State": {
          "$ref": "#/definitions/CommandState"
        },
        "TaskID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "CommandState": {
      "type": "integer",
      "format": "uint8",
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
...
    "GenericIdResponse": {
      "type": "object",
      "properties": {
        "ID": {
          "description": "Id\nrequired",
          "type": "string",
          "x-go-name": "Id"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/api"
    },
...
    "Job": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Label": {
          "type": "string"
        },
        "QueueID": {
          "type": "integer",
          "format": "uint64"
        },
        "State": {
          "$ref": "#/definitions/JobState"
        },
        "Tasks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Task"
          }
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "JobState": {
      "type": "integer",
      "format": "uint8",
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
...
    "Queue": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Jobs": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Job"
          }
        },
        "Name": {
          "type": "string"
        },
        "Nodes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ResourceNode"
          }
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "QueueResponse": {
      "type": "object",
      "properties": {
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Name": {
          "type": "string"
        },
        "Nodes": {
          "type": "integer",
          "format": "uint64"
        },
        "PendingTasks": {
          "type": "integer",
          "format": "uint64"
        },
        "RunningTasks": {
          "type": "integer",
          "format": "uint64"
        },
        "Workers": {
          "type": "integer",
          "format": "uint64"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/api"
    },
    "ResourceNode": {
      "type": "object",
      "properties": {
        "Address": {
          "type": "string"
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "QueueID": {
          "type": "integer",
          "format": "uint64"
        },
        "State": {
          "$ref": "#/definitions/ResourceState"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "ResourceState": {
      "type": "integer",
      "format": "uint8",
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
...
    "Task": {
      "type": "object",
      "properties": {
        "Commands": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Command"
          }
        },
        "Config": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskConfig"
          }
        },
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "JobID": {
          "type": "integer",
          "format": "uint64"
        },
        "Metadata": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskMetadata"
          }
        },
        "State": {
          "$ref": "#/definitions/TaskState"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "TaskConfig": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Key": {
          "type": "string"
        },
        "TaskID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Value": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "TaskMetadata": {
      "type": "object",
      "properties": {
        "CreatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "DeletedAt": {
          "type": "string",
          "format": "date-time"
        },
        "ID": {
          "type": "integer",
          "format": "uint64"
        },
        "Key": {
          "type": "string"
        },
        "TaskID": {
          "type": "integer",
          "format": "uint64"
        },
        "UpdatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "Value": {
          "type": "string"
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
    "TaskSpec": {
      "type": "object",
      "properties": {
        "Commands": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Config": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "ID": {
          "type": "string"
        },
        "Metadata": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        }
      },
      "x-go-package": "github.com/emanueljoivo/arrebol/api"
    },
    "TaskState": {
      "type": "integer",
      "format": "uint8",
      "x-go-package": "github.com/emanueljoivo/arrebol/storage"
    },
...
    "jobSpec": {
      "type": "object",
      "required": [
        "Label",
        "Tasks"
      ],
      "properties": {
        "Label": {
          "description": "label",
          "type": "string"
        },
        "Tasks": {
          "description": "tasks",
          "type": "array",
          "items": {
            "$ref": "#/definitions/TaskSpec"
          }
        }
      },
      "x-go-name": "JobSpec",
      "x-go-package": "github.com/emanueljoivo/arrebol/api"
    }
  }
}